refactor(ai-worker): alinhar imageplanning ao padrão wireframe
